# Customer/Program.cs
using Microsoft.Extensions.Caching.Distributed;

var builder = WebApplication.CreateBuilder(args);

// Aspire defaults: telemetry, health checks, service discovery, HttpClient resilience.
builder.AddServiceDefaults();

// One MongoDB client per process, shared by the repositories and the EventStore. Aspire injects ConnectionStrings:mongodb; the resolver also accepts every legacy
// shape, and fails with a message naming each key it tried rather than a null-argument throw.
builder.Services.AddSingleton<IMongoClient>(_ =>
    new MongoClient(MongoConnectionResolver.Resolve(builder.Configuration)));

// Readiness probe. Singleton so the 5s result cache is process-wide.
builder.Services.AddSingleton<MongoHealthCheck>();
builder.Services.AddHealthChecks()
    .AddCheck<MongoHealthCheck>("mongodb", tags: ["ready"]);

// Add MongoDB
builder.Services.AddMongoDbRepository(builder.Configuration);

// Add Distributed Cache
builder.Services.AddDistributedMemoryCache();

// Add MediatR
builder.Services.AddMediatR(cfg => { cfg.RegisterServicesFromAssembly(typeof(Program).Assembly); });
builder.Services.AddEventStore();

// Add services required to support using MVC's model binders
builder.Services.AddMvcCore();

// F-014: ObjectId has no JSON representation of its own, so System.Text.Json serialises the struct's public
// properties and emits `"id": { "timestamp": …, "machine": … }` — a shape that cannot be read back into an
// ObjectId at all. Three of F-014's route families need the id from a create response in order to work
// (PUT /notes/{id}, POST /messages/{id}/read, POST /notifications/{id}/read), so this is load-bearing rather
// than cosmetic. Pre-existing for every other route that returns an entity; see ObjectIdJsonConverter.
builder.Services.ConfigureHttpJsonOptions(options =>
    options.SerializerOptions.Converters.Add(new ObjectIdJsonConverter()));

// Register Singleton instances
builder.Services.AddSingleton<IKafkaClient, KafkaClient>();
// Scoped, not Singleton: RequestCollection consumes the scoped IEventStore, and a
// singleton capturing it fails DI validation — which is enabled in Development, the
// environment the AppHost runs services in. RequestCollection is stateless, so request
// scope is the correct lifetime rather than a workaround.
builder.Services.AddScoped<IRequestCollection, RequestCollection>();

// Enable & configure JSON Problem Details error responses
// ADR-022 / F-016-T08: ForbiddenException -> 403 centrally, so an endpoint that omits a local
// try/catch returns 403 rather than a bare 500. Registered unconditionally, unlike the
// Development-only UseExceptionHandler lambda below.
builder.Services.AddExceptionHandler<AgendaBuddyExceptionHandler>();
builder.Services.AddProblemDetails(options =>
    options.CustomizeProblemDetails = context => CustomizeProblemDetails(context.ProblemDetails, context.HttpContext));

// Add Anti-CSRF/XSRF services
builder.Services.AddAntiforgery();

// JWT Bearer authentication (reads JWT_PUBLIC_KEY env var — fails fast if absent)
builder.Services.AddAgendaBuddyAuthentication();
builder.Services.AddAuthorization();

builder.Services.AddEndpointsApiExplorer();
builder.Services.AddSwaggerGen();
var app = builder.Build();

// /health runs every check; /alive only the live-tagged ones, so a service waiting on MongoDB is
// not restarted for being unready.
app.MapDefaultEndpoints();

if (app.Environment.IsDevelopment())
{
    app.UseSwagger();
    app.UseSwaggerUI();
    // Error handling
    app.UseExceptionHandler(new ExceptionHandlerOptions
    {
        AllowStatusCode404Response = true,
        ExceptionHandler = async exceptionContext =>
        {
            // GitHub issue to support this in framework: https://github.com/dotnet/aspnetcore/issues/43831
            var exceptionHandlerFeature = exceptionContext.Features.Get<IExceptionHandlerFeature>();

            if (exceptionHandlerFeature?.Error is BadHttpRequestException badRequestEx)
                exceptionContext.Response.StatusCode = badRequestEx.StatusCode;

            if (exceptionContext.Request.AcceptsJson()
                && exceptionContext.RequestServices.GetRequiredService<IProblemDetailsService>() is
                { } problemDetailsService)
            {
                // Write as JSON problem details
                await problemDetailsService.WriteAsync(new ProblemDetailsContext
                {
                    HttpContext = exceptionContext,
                    AdditionalMetadata = exceptionHandlerFeature?.Endpoint?.Metadata,
                    ProblemDetails = { Status = exceptionContext.Response.StatusCode }
                });
            }
            else
            {
                exceptionContext.Response.ContentType = "text/plain";
                var message = ReasonPhrases.GetReasonPhrase(exceptionContext.Response.StatusCode) switch
                {
                    { Length: > 0 } reasonPhrase => reasonPhrase,
                    _ => "An error occurred"
                };
                await exceptionContext.Response.WriteAsync(message + "\r\n");
                await exceptionContext.Response.WriteAsync(
                    $"Request ID: {Activity.Current?.Id ?? exceptionContext.TraceIdentifier}");
            }
        }
    });
}

// MUST stay AFTER the IsDevelopment() block. Middleware registered earlier is outermost and an
// exception propagates outward, so the INNERMOST handler sees it first. Placed here, this one takes
// ForbiddenException and declines everything else, which then rethrows and reaches the Development
// lambda exactly as it does today. Placed BEFORE that block, the lambda would swallow
// ForbiddenException and the central 403 would fail in Development only. See AgendaBuddyExceptionHandler.
app.UseExceptionHandler();

// F-021 PRD requirement 13: HSTS (under its flag) and the HTTPS redirect run BEFORE authentication.
// Registered after UseAuthentication, as it was until F-021, the redirect parsed and validated the
// bearer token out of a plaintext request and only then told the client to come back over TLS.
app.UseAgendaBuddyTransportSecurity();

app.UseAntiforgery();
app.UseAuthentication();
app.UseAuthorization();
app.UseStatusCodePages();

var customers = app.MapGroup("/api/v1/customers")
    .WithTags("CustomerAPI")
    .WithOpenApi()
    .AddEndpointFilter<ProblemDetailsServiceEndpointFilter>();

customers.MapPost("/", async Task<Results<ValidationProblem, Created<CustomerEntity>>> (
    IMediator mediator,
    CustomerService customerService,
    CustomerEntity customerEntity,
    IRequestCollection requestCollection) =>
{
    if (!MiniValidator.TryValidate(customerEntity, out var errors))
        return TypedResults.ValidationProblem(errors);
    var filter =
        SupportTools<CustomerEntity>.FilterByNameAndLastName(customerEntity.FirstName!, customerEntity.LastName!);
    var existingCustomer = await customerService.FindCustomerAsync(filter);
    var topicName = KafkaHelper.CreateCustomerTopicName(customerEntity.Email!);
    if (existingCustomer != null)
        return TypedResults.ValidationProblem(GenerateErrorMessage(
            "Existing record found", new[]
            {
                $"Email:{customerEntity.Email}"
            }));

    var eventResponse =
        await EventsHelper.AddCustomerEvent(requestCollection, mediator, customerService, customerEntity);
    if (!string.IsNullOrEmpty(eventResponse) && !eventResponse.ToLower().StartsWith("exception"))
        return TypedResults.Created($"/api/v1/customers/{customerEntity.Id}", customerEntity);

    return TypedResults.ValidationProblem(GenerateErrorMessage(
        "Kafka Error", new[] { "Kafka Topic", $"{topicName}" })
    );
})
.WithName("CreateCustomer")
.RequireAuthorization();

customers.MapPut("/{email}",
    async Task<Results<ValidationProblem, ForbidHttpResult, NotFound, Accepted>> (string email,
        ClaimsPrincipal user,
        IMediator mediator,
        CustomerService customerService, CustomerEntity customerEntity, IRequestCollection requestCollection) =>
    {
        if (!MiniValidator.TryValidate(customerEntity, out var errors))
            return TypedResults.ValidationProblem(errors);

        // Deliberately NOT wrapped in try/catch — F-016 AC-13. This is the route that demonstrates the
        // central mapping: AgendaBuddyExceptionHandler turns ForbiddenException into 403 whether or not
        // an endpoint remembered to catch it. Before F-016 this line without a catch produced a 500 (and
        // in Production, a bare empty-bodied one). Removing the catch here rather than shipping a
        // test-only endpoint also demonstrates AC-14's no-double-handling in the same stroke.
        // ForbidHttpResult stays in the union above on purpose: this route still returns 403, so removing
        // it would drop 403 from the generated OpenAPI while the behaviour was unchanged.
        OwnershipGuard.AssertOwner(user, email);

        var eventResponse =
            await EventsHelper.UpdateCustomerEvent(email, requestCollection, mediator, customerService, customerEntity);

        if (!string.IsNullOrEmpty(eventResponse)) return TypedResults.Accepted("api/v1/customers");

        return TypedResults.NotFound();
    })
    .WithName("UpdateCustomer")
    .RequireAuthorization();

customers.MapGet("",
    async Task<Ok<PagedResponse<CustomerEntity>>> (IMediator mediator,
        ClaimsPrincipal user,
        CustomerService customerService, IRequestCollection requestCollection, IDistributedCache cache,
        int? page = null, int? pageSize = null) =>
    {
        // F-016 AC-22 / threat T-003 / ADR-026: the Provider role, not merely a token. Authenticating this
        // route alone was nearly worthless -- POST /api/v1/auth/register is anonymous, unverified and
        // unrate-limited, so an attacker self-registers as a Customer and pages the whole customer table
        // exactly as before. Pagination bounds the response, not the extraction.
        //
        // No local try/catch: T08's AgendaBuddyExceptionHandler maps ForbiddenException to 403 centrally.
        // Guard runs BEFORE the cache read, so a refused caller never reaches cached data.
        OwnershipGuard.AssertRole(user, "Provider");

        // F-016 AC-15 / ADR-023. See Provider/Program.cs for why clamping rather than rejecting.
        var pageRequest = PageRequest.Clamp(page, pageSize);

        // ⚠️ The cache key carries the page, or page 2 would serve page 1's entry. Cheap to get wrong and
        // invisible in a single-page test.
        var key = $"customers-p{pageRequest.Page}-s{pageRequest.PageSize}";
        var customerCollection = await cache.GetOrCreateAsync(key,
            async token => await EventsHelper.GetCustomersEvent(
                requestCollection, mediator, customerService, pageRequest));

        // 204 is RETIRED (ADR-023).
        return customerCollection is not null
            ? TypedResults.Ok(customerCollection)
            : TypedResults.Ok(PagedResponse<CustomerEntity>.From([], 0, pageRequest));
    })
    // F-016 AC-8 / requirement 9: PII-bearing read, so no longer anonymous. Breaking change with zero reachable consumers (01-api-surface.md:158).
    .WithName("GetAllCustomers")
    .RequireAuthorization();

customers.MapGet("/{email}", async Task<Results<Ok<CustomerEntity>, NotFound>> (IMediator mediator, string email,
    CustomerService customerService, IRequestCollection requestCollection, IDistributedCache cache) =>
{
    var key = $"customers-{email}";

    var customer = await cache.GetOrCreateAsync(key,
        async token => await EventsHelper.GetCustomerByEmailEvent(requestCollection, mediator, customerService, email));

    if (customer is not null)
        return TypedResults.Ok(customer);

    return TypedResults.NotFound();
})
    // F-016 AC-8 / requirement 9: PII-bearing read, so no longer anonymous. Breaking change with zero reachable consumers (01-api-surface.md:158).
    .WithName("GetCustomerByEmail")
    .RequireAuthorization();

// ── F-014: messages and notifications ────────────────────────────────────────────────────────────────
//
// TWO NEW TOP-LEVEL ROUTE GROUPS in this process, not children of /api/v1/customers — and that is the point
// (ADR D-2). A message is addressed to a PERSON: a provider has an inbox for exactly the same reason a
// customer does, so a URL saying `customers` about a provider's inbox would assert something false and every
// client would have to work around it. Identity already hosts two unrelated groups (`/api/v1/auth` and
// `/device-token`), so this is a precedent rather than a novelty. The Customer service hosts them because it
// already owns the provider↔customer relationship these messages travel along.

var messages = app.MapGroup("/api/v1/messages")
    .WithTags("MessageAPI")
    .WithOpenApi()
    .AddEndpointFilter<ProblemDetailsServiceEndpointFilter>();

// Threat T-204: the recipient is the caller's `sub` claim and there is NO parameter. A recipient parameter
// would be a thing to tamper with — `MessageService.GetInboxAsync` takes one, and passing a client-supplied
// value through would hand any authenticated caller anyone else's inbox.
messages.MapGet("/", async Task<Results<Ok<IEnumerable<MessageEntity>>, ForbidHttpResult>> (
        ClaimsPrincipal user, IMessageService service) =>
    {
        var caller = user.FindFirstValue(ClaimTypes.NameIdentifier);
        if (caller is null) return TypedResults.Forbid();

        return TypedResults.Ok(await service.GetInboxAsync(caller));
    })
    .WithName("GetInbox")
    .RequireAuthorization();

// Threat T-204: ONE counterpart in the URL. `MessageService` derives thread_id by sorting both addresses, so
// with the caller always supplying one side, a thread between two other people has no representation in this
// URL space at all — it is unrequestable rather than merely refused.
messages.MapGet("/thread/{counterpartEmail}",
        async Task<Results<Ok<IEnumerable<MessageEntity>>, ForbidHttpResult>> (
            string counterpartEmail, ClaimsPrincipal user, IMessageService service) =>
        {
            var caller = user.FindFirstValue(ClaimTypes.NameIdentifier);
            if (caller is null) return TypedResults.Forbid();

            return TypedResults.Ok(await service.GetThreadAsync(caller, counterpartEmail));
        })
    .WithName("GetMessageThread")
    .RequireAuthorization();

messages.MapPost("/", async Task<Results<Created<MessageEntity>, ForbidHttpResult, BadRequest<string>>> (
        MessageRequest request, ClaimsPrincipal user, IMessageService service) =>
    {
        if (request is null || string.IsNullOrWhiteSpace(request.RecipientEmail)
                            || string.IsNullOrWhiteSpace(request.Body))
            return TypedResults.BadRequest("recipientEmail and body are required.");

        var caller = user.FindFirstValue(ClaimTypes.NameIdentifier);
        if (caller is null) return TypedResults.Forbid();

        // The sender is the caller. MessageRequest has no sender field, which is the cheapest guarantee that
        // no future refactor trusts one from the body.
        var message = new MessageEntity
        {
            SenderEmail = caller,
            RecipientEmail = request.RecipientEmail,
            Body = request.Body
        };

        await service.SendMessageAsync(message);
        return TypedResults.Created($"/api/v1/messages/{message.Id}", message);
    })
    .WithName("SendMessage")
    .RequireAuthorization();

// Only the RECIPIENT may mark a message read. A sender marking their own message read is meaningless, and
// permitting it would let a sender probe whether an id exists.
messages.MapPost("/{id}/read", async Task<Results<NoContent, ForbidHttpResult>> (
        string id, ClaimsPrincipal user, IMessageService service, IRepository<MessageEntity> repository) =>
    {
        var caller = user.FindFirstValue(ClaimTypes.NameIdentifier);
        var message = await repository.FindOneAsync(new BsonDocument("_id", new ObjectId(id)));

        // A missing message and someone else's answer identically — the same rule the notes routes follow, so
        // this cannot be used to enumerate message ids.
        try { OwnershipGuard.AssertOwner(user, message?.RecipientEmail); }
        catch (ForbiddenException) { return TypedResults.Forbid(); }

        await service.MarkReadAsync(id);
        return TypedResults.NoContent();
    })
    .WithName("MarkMessageRead")
    .RequireAuthorization();

var notifications = app.MapGroup("/api/v1/notifications")
    .WithTags("NotificationAPI")
    .WithOpenApi()
    .AddEndpointFilter<ProblemDetailsServiceEndpointFilter>();

// ⚠️ THERE IS DELIBERATELY NO ROUTE THAT CREATES A NOTIFICATION (threat T-208). Notifications are produced by
// domain events, not by users: a create route would let any authenticated caller write a convincing "Your
// appointment was cancelled" into somebody else's list. `NotificationService.SendAsync` stays reachable
// in-process to whatever writes one.
//
// The consequence, stated so an empty list is not read as a bug: NOTHING WRITES A NOTIFICATION YET. No domain
// event calls SendAsync, so this route returns [] until something does. F-014 requirement 19 — storage
// without delivery, and for now without production either.
notifications.MapGet("/", async Task<Results<Ok<IEnumerable<NotificationEntity>>, ForbidHttpResult>> (
        ClaimsPrincipal user, INotificationService service) =>
    {
        var caller = user.FindFirstValue(ClaimTypes.NameIdentifier);
        if (caller is null) return TypedResults.Forbid();

        return TypedResults.Ok(await service.GetForRecipientAsync(caller));
    })
    .WithName("GetNotifications")
    .RequireAuthorization();

notifications.MapPost("/{id}/read", async Task<Results<NoContent, ForbidHttpResult>> (
        string id, ClaimsPrincipal user, INotificationService service, IRepository<NotificationEntity> repository) =>
    {
        var notification = await repository.FindOneAsync(new BsonDocument("_id", new ObjectId(id)));

        try { OwnershipGuard.AssertOwner(user, notification?.RecipientEmail); }
        catch (ForbiddenException) { return TypedResults.Forbid(); }

        await service.MarkReadAsync(id);
        return TypedResults.NoContent();
    })
    .WithName("MarkNotificationRead")
    .RequireAuthorization();

app.Run();

// Functions and Methods
void CustomizeProblemDetails(ProblemDetails problemDetails, HttpContext httpContext)
{
    problemDetails.Extensions["requestId"] = Activity.Current?.Id ?? httpContext.TraceIdentifier;
}

Dictionary<string, string[]> GenerateErrorMessage(string key, string[] values)
{
    var dictionary = new Dictionary<string, string[]> { { key, values } };
    return dictionary;
}